+++ package.json
@@ -57,7 +57,6 @@
     "not ie <= 10"
   ],
   "dependencies": {
-    "@ant-design/compatible": "^0.0.1-rc.0",
     "@ant-design/icons": "^4.0.0-alpha.19",
     "@ant-design/pro-layout": "4.11.0-1",
     "@ant-design/pro-table": "^1.0.16",
